+++ packages/signers/pjs-signer/src/pjs-signed-extensions-mappers.ts
@@ -14,9 +14,11 @@ type SignedExtension = {
   additionalSigned: Uint8Array
 }
 
-const toPjsHex = (value: number | bigint) => {
-  const inner = value.toString(16)
-  return "0x" + (inner.length % 2 ? "0" : "") + inner
+const toPjsHex = (value: number | bigint, minByteLen?: number) => {
+  let inner = value.toString(16)
+  inner = (inner.length % 2 ? "0" : "") + inner
+  const nPaddedBytes = Math.max(0, (minByteLen || 0) - inner.length / 2)
+  return "0x" + "00".repeat(nPaddedBytes) + inner
 }
 
 export const CheckGenesis = ({
@@ -28,13 +30,14 @@ export const CheckGenesis = ({
 export const CheckNonce = ({
   value,
 }: SignedExtension): { nonce: HexString } => {
-  return { nonce: toPjsHex(compact.dec(value)) }
+  // nonce is a u32 in pjs => 4 bytes
+  return { nonce: toPjsHex(compact.dec(value), 4) }
 }
 
 export const CheckTxVersion = ({
   additionalSigned,
 }: SignedExtension): { transactionVersion: HexString } => {
-  return { transactionVersion: toPjsHex(u32.dec(additionalSigned)) }
+  return { transactionVersion: toPjsHex(u32.dec(additionalSigned), 4) }
 }
 
 const assetTxPaymentDec = Struct({
@@ -49,14 +52,14 @@ export const ChargeAssetTxPayment = ({
 
   return {
     ...(asset ? { assetId: toHex(asset) } : {}),
-    ...(tip > 0n ? { tip: toPjsHex(tip) } : {}),
+    tip: toPjsHex(tip, 16),
   }
 }
 
 export const ChargeTransactionPayment = ({
   value,
 }: SignedExtension): { tip: HexString } => ({
-  tip: toPjsHex(compactBn.dec(value)),
+  tip: toPjsHex(compactBn.dec(value), 16), // u128 => 16 bytes
 })
 
 export const CheckMortality = (
@@ -65,11 +68,11 @@ export const CheckMortality = (
 ): { era: HexString; blockHash: HexString; blockNumber: HexString } => ({
   era: toHex(value),
   blockHash: toHex(additionalSigned),
-  blockNumber: toPjsHex(blockNumber),
+  blockNumber: toPjsHex(blockNumber, 4),
 })
 
 export const CheckSpecVersion = ({
   additionalSigned,
 }: SignedExtension): { specVersion: HexString } => ({
-  specVersion: toPjsHex(u32.dec(additionalSigned)),
+  specVersion: toPjsHex(u32.dec(additionalSigned), 4),
 })